Boring cleanup

+++ include/3dstris/util.hpp
@@ -14,11 +14,11 @@ std::unique_ptr<T> make_unique(Args&&... args) {
 	return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
 }
 
-constexpr u16 SCREEN_WIDTH = 400;
-constexpr u16 SCREEN_HEIGHT = 240;
+constexpr u32 SCREEN_WIDTH = 400;
+constexpr u32 SCREEN_HEIGHT = 240;
 
-constexpr u16 BSCREEN_WIDTH = 320;
-constexpr u16 BSCREEN_HEIGHT = 240;
+constexpr u32 BSCREEN_WIDTH = 320;
+constexpr u32 BSCREEN_HEIGHT = 240;
 
 using Color = u32;
 inline void straightLine(const Pos pos, const WH wh, const float thick,
